+++ database/migrations/2019_05_14_095812_add_id_column_to_dn_lifestyle_sterile_table.php
@@ -14,7 +14,7 @@ class AddIdColumnToDnLifestyleSterileTable extends Migration
     public function up()
     {
         Schema::table('dn_lifestyle_sterile', function (Blueprint $table) {
-            $table->increments('id')->first();
+            $table->unsignedTinyInteger('id')->first();
         });
     }
