+++ database/migrations/2019_05_14_102409_add_id_column_to_dn_s2_allergy_list_table.php
@@ -15,7 +15,7 @@ class AddIdColumnToDnS2AllergyListTable extends Migration
     {
         Schema::table('dn_s2_allergy_list', function (Blueprint $table) {
 
-            $table->increments('id')->first();
+            $table->unsignedTinyInteger('id')->first();
 
         });
     }
