+++ database/migrations/2019_05_14_111033_add_id_column_to_dn_s3_neurologic_ab_table.php
@@ -15,7 +15,7 @@ class AddIdColumnToDnS3NeurologicAbTable extends Migration
     {
         Schema::table('dn_s3_neurologic_ab', function (Blueprint $table) {
 
-            $table->increments('id')->first();
+            $table->unsignedTinyInteger('id')->first();
 
         });
     }
